virtcontainers: change uint32 to uint64 for ioctl
The PR changes the parameter args from uint32 to uint64 for ioctl function.
That leads to an endianess bug.

Fixes: #947

+++ virtcontainers/utils/utils_linux_test.go
@@ -15,7 +15,7 @@ import (
 func TestFindContextID(t *testing.T) {
 	assert := assert.New(t)
 
-	ioctlFunc = func(fd uintptr, request int, arg1 uint32) error {
+	ioctlFunc = func(fd uintptr, request int, arg1 uint64) error {
 		return errors.New("ioctl")
 	}
